Elevate Graph objects from Elements sub folder to own primary folder for Analytical oM #1261
