+++ Analytical_oM/Graph/Graph.cs
@@ -21,8 +21,6 @@
  */
 
 using BH.oM.Base;
-using BH.oM.Geometry;
-using BH.oM.Graphics.Views;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -30,12 +28,21 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Analytical.Elements
+namespace BH.oM.Analytical.Graph
 {
-    [Description("Graphical projection for Graphs.")]
-    public class GraphicalProjection : BHoMObject, IProjection
+    [Description("A set of IBHoMObject entities and the relations between them.")]
+    public class Graph : BHoMObject
     {
-        public virtual IView View { get; set; } = null;
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
+
+        [Description("Dictionary of IBHoMObject entities.")]
+        public virtual Dictionary<Guid, IBHoMObject> Entities { get; set; } = new Dictionary<Guid, IBHoMObject>();
+
+        [Description("Collection of relations between IBHoMObject entities.")]
+        public virtual List<IRelation> Relations { get; set; } = new List<IRelation>();
+
     }
 }
