+++ Analytical_oM/Graph/Projection/IProjection.cs
@@ -28,12 +28,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Analytical.Elements
+namespace BH.oM.Analytical.Graph.Projection
 {
-    [Description("Spatial Projection for Graphs.")]
-    public class SpatialProjection : BHoMObject, IProjection
+    [Description("Interface common to all graph projections.")]
+    public interface IProjection : IBHoMObject
     {
-       
+
     }
 }
